unrestrict helpers
I plan to unrestrict a container of CoEdges, but I can't give a DataL
instance for CoEdge or the user of the library will be able to
unrestrict them. So I'll have a private unrestrictCoEdge function, and
I'll need some helpers to turn it into an unrestrict function for the
containers since I won't be able to rely on instance search.